requirements.txt
npyscreen
GeoIP
python-magic
img2txt.py
